avoid cyclical dependencies by splitting off config type in package config; move module config logic to core
